+++ programs/artiste/src/vsx_widget/controllers/vsx_widget_controller_color.h
@@ -46,7 +46,7 @@ public:
   void event_mouse_up(vsx_widget_distance distance,vsx_widget_coords coords,int button);
   void event_mouse_double_click(vsx_widget_distance distance,vsx_widget_coords coords,int button);
   void event_mouse_move(vsx_widget_distance distance,vsx_widget_coords coords);
-  bool event_key_down(signed long key);
+  bool event_key_down(uint16_t key);
   void command_process_back_queue(vsx_command_s *t);
 
   vsx_widget_controller_color();
