+++ programs/artiste/src/vsx_widget/sequencer/pool_manager/vsx_widget_seq_pool.h
@@ -57,7 +57,7 @@ public:
   void init();
   void i_draw();
   void event_text(wchar_t character_wide, char character);
-  bool event_key_down(signed long key);
+  bool event_key_down(uint16_t key);
   void set_server(vsx_widget* serv);
   vsx_widget_seq_pool_manager();
 };
